refactor(integration-card): moved component to  folder
