+++ frontend/src/lib/components/integration/integration-card.svelte
